Please enter the commit message for your changes. Lines starting *removed client code from function waitQueve() and placed in seperate file *made header file to access tcp_client calls *now can easily send and recieve data from the client by including the header "tcp_client.h" *fixed some bugs with messageScreen in gamePlay.c *dried up code in gamePlay.c
