--- mock_up_networking/READ_ME.txt
@@ -1,3 +1,0 @@
-These files are a basic implementation of the server client model.
-
-They are currently implemented to run on flip3.
