+++ tcp_client.h
@@ -1,0 +1,15 @@
+#ifndef TCP_CLIENT_H
+#define TCP_CLIENT_H
+
+#include <stdio.h>
+#include <stdlib.h>
+#include <sys/types.h>
+#include <sys/socket.h>
+#include <netinet/in.h>
+#include <arpa/inet.h>
+
+int establish_connection();
+int receive_data();
+void closing_connection();
+
+#endif
